+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+- Convertion to size in a readable form with binary prefix (e.g. KiB, etc.)
+
 ## [0.0.2] (2019-01-25)
 
 ### Added
@@ -21,5 +26,6 @@ Display a hover with following conversions of the pointed hexadecimal value:
 
 
 
+[Unreleased]: https://github.com/mateuszchudyk/vscode-hexinspector/compare/v0.0.2...HEAD
 [0.0.2]: https://github.com/mateuszchudyk/vscode-hexinspector/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/mateuszchudyk/vscode-hexinspector/commit/165f3ed7caeed85a803346ce4e36781e25abb1ce