+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Convertion to size in a readable form with binary prefix (e.g. KiB, etc.)
+- Display length of a hexadecimal value in bytes
 
 ## [0.0.2] (2019-01-25)