+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 - Make bytesToFloat32 compatible with IEEE 754
 - Make bytesToFloat64 compatible with IEEE 754
+- Fix bug in bytesToSignedDec (it unset the highest bit of the input)
 
 ## [0.0.2] (2019-01-25)